server_app/requirements.txt
djangorestframework
djangorestframework-authtoken
python-dotenv
psycopg2
requests